Refactor model/src/conditions and improve type annotations
