+++ model/src/conditions/index.ts
@@ -7,15 +7,6 @@ export {
 } from "./inline-condition-operators";
 
 export {
-  Condition,
-  ConditionsModel,
-  Field,
-  GroupDef,
-  ConditionRef,
-  toPresentationString,
-} from "./inline-condition-model";
-
-export {
   ConditionValue,
   timeUnits,
   valueFrom,
@@ -24,3 +15,10 @@ export {
   dateTimeUnits,
   dateUnits,
 } from "./inline-condition-values";
+
+export { Field } from "./field";
+export { GroupDef } from "./group-def";
+export { Condition } from "./condition";
+export { ConditionRef } from "./condition-ref";
+export { ConditionsModel } from "./inline-condition-model";
+export { toExpression, toPresentationString } from "./helpers";
