+++ model/src/conditions/inline-condition-model.ts
@@ -1,21 +1,17 @@
-import ComponentTypes from "../component-types";
-import { getExpression } from "./inline-condition-operators";
-import { AbstractConditionValue, valueFrom } from "./inline-condition-values";
+import { Field } from "./field";
+import { GroupDef } from "./group-def";
+import { Condition } from "./condition";
+import { ConditionRef } from "./condition-ref";
+import { ConditionGroup } from "./condition-group";
+import { valueFrom } from "./inline-condition-values";
+import { Coordinator, toPresentationString, toExpression } from "./helpers";
 
-export const coordinators = {
-  AND: "and",
-  OR: "or",
-};
+type ConditionsArray = (Condition | ConditionGroup | ConditionRef)[];
 
 export class ConditionsModel {
-  #groupedConditions;
-  #userGroupedConditions;
-  #conditionName;
-
-  constructor() {
-    this.#groupedConditions = [];
-    this.#userGroupedConditions = [];
-  }
+  #groupedConditions: ConditionsArray = [];
+  #userGroupedConditions: ConditionsArray = [];
+  #conditionName: string | undefined = undefined;
 
   clone() {
     const toReturn = new ConditionsModel();
@@ -44,20 +40,24 @@ export class ConditionsModel {
     return this.#conditionName;
   }
 
-  add(condition) {
+  add(condition: Condition) {
     const coordinatorExpected = this.#userGroupedConditions.length !== 0;
+
     if (condition.getCoordinator() && !coordinatorExpected) {
       throw Error("No coordinator allowed on the first condition");
     } else if (!condition.getCoordinator() && coordinatorExpected) {
       throw Error("Coordinator must be present on subsequent conditions");
     }
+
     this.#userGroupedConditions.push(condition);
     this.#groupedConditions = this._applyGroups(this.#userGroupedConditions);
+
     return this;
   }
 
-  replace(index, condition) {
+  replace(index: number, condition: Condition) {
     const coordinatorExpected = index !== 0;
+
     if (condition.getCoordinator() && !coordinatorExpected) {
       throw Error("No coordinator allowed on the first condition");
     } else if (!condition.getCoordinator() && coordinatorExpected) {
@@ -67,12 +67,14 @@ export class ConditionsModel {
         `Cannot replace condition ${index} as no such condition exists`
       );
     }
+
     this.#userGroupedConditions.splice(index, 1, condition);
     this.#groupedConditions = this._applyGroups(this.#userGroupedConditions);
+
     return this;
   }
 
-  remove(indexes) {
+  remove(indexes: number[]) {
     this.#userGroupedConditions = this.#userGroupedConditions
       .filter((_condition, index) => !indexes.includes(index))
       .map((condition, index) =>
@@ -83,7 +85,7 @@ export class ConditionsModel {
     return this;
   }
 
-  addGroups(groupDefs) {
+  addGroups(groupDefs: GroupDef[]) {
     this.#userGroupedConditions = this._group(
       this.#userGroupedConditions,
       groupDefs
@@ -92,7 +94,7 @@ export class ConditionsModel {
     return this;
   }
 
-  splitGroup(index) {
+  splitGroup(index: number) {
     this.#userGroupedConditions = this._ungroup(
       this.#userGroupedConditions,
       index
@@ -101,7 +103,7 @@ export class ConditionsModel {
     return this;
   }
 
-  moveEarlier(index) {
+  moveEarlier(index: number) {
     if (index > 0 && index < this.#userGroupedConditions.length) {
       this.#userGroupedConditions.splice(
         index - 1,
@@ -116,7 +118,7 @@ export class ConditionsModel {
     return this;
   }
 
-  moveLater(index) {
+  moveLater(index: number) {
     if (index >= 0 && index < this.#userGroupedConditions.length - 1) {
       this.#userGroupedConditions.splice(
         index + 1,
@@ -162,7 +164,9 @@ export class ConditionsModel {
       .join(" ");
   }
 
-  _applyGroups(userGroupedConditions) {
+  _applyGroups(
+    userGroupedConditions: (Condition | ConditionGroup | ConditionRef)[]
+  ) {
     const correctedUserGroups = userGroupedConditions.map((condition) =>
       condition instanceof ConditionGroup && condition.conditions.length > 2
         ? new ConditionGroup(
@@ -173,15 +177,17 @@ export class ConditionsModel {
           )
         : condition
     );
+
     return this._group(
       correctedUserGroups,
       this._autoGroupDefs(correctedUserGroups)
     );
   }
 
-  _group(conditions, groupDefs) {
+  _group(conditions: ConditionsArray, groupDefs: GroupDef[]) {
     return conditions.reduce((groups, condition, index, conditions) => {
       const groupDef = groupDefs.find((groupDef) => groupDef.contains(index));
+
       if (groupDef) {
         if (groupDef.startsWith(index)) {
           const groupConditions = groupDef.applyTo(conditions);
@@ -190,11 +196,12 @@ export class ConditionsModel {
       } else {
         groups.push(condition);
       }
+
       return groups;
-    }, []);
+    }, [] as ConditionsArray);
   }
 
-  _ungroup(conditions, splitIndex) {
+  _ungroup(conditions: ConditionsArray, splitIndex: number) {
     if (conditions[splitIndex].isGroup()) {
       const copy = [...conditions];
       copy.splice(
@@ -207,17 +214,20 @@ export class ConditionsModel {
     return conditions;
   }
 
-  _autoGroupDefs(conditions) {
+  _autoGroupDefs(conditions: ConditionsArray) {
     const orPositions: number[] = [];
+
     conditions.forEach((condition, index) => {
-      if (condition.getCoordinator() === coordinators.OR) {
+      if (condition.getCoordinator() === Coordinator.OR) {
         orPositions.push(index);
       }
     });
-    const hasAnd = !!conditions.find(
-      (condition) => condition.getCoordinator() === coordinators.AND
-    );
+
     const hasOr = orPositions.length > 0;
+    const hasAnd = !!conditions.find(
+      (condition) => condition.getCoordinator() === Coordinator.AND
+    );
+
     if (hasAnd && hasOr) {
       let start = 0;
       const groupDefs: GroupDef[] = [];
@@ -234,6 +244,7 @@ export class ConditionsModel {
       });
       return groupDefs;
     }
+
     return [];
   }
 
@@ -246,7 +257,9 @@ export class ConditionsModel {
     };
   }
 
-  static from(obj) {
+  static from(
+    obj: ConditionsModel | { name: string; conditions: Condition[] }
+  ) {
     if (obj instanceof ConditionsModel) {
       return obj;
     }
@@ -281,272 +294,4 @@ function conditionFrom(it) {
     valueFrom(it.value),
     it.coordinator
   );
-}
-
-export class GroupDef {
-  first;
-  last;
-
-  constructor(first, last) {
-    if (typeof first !== "number" || typeof last !== "number") {
-      throw Error(`Cannot construct a group from ${first} and ${last}`);
-    } else if (first >= last) {
-      throw Error(`Last (${last}) must be greater than first (${first})`);
-    }
-    this.first = first;
-    this.last = last;
-  }
-
-  contains(index) {
-    return this.first <= index && this.last >= index;
-  }
-
-  startsWith(index) {
-    return this.first === index;
-  }
-
-  applyTo(conditions) {
-    return [...conditions].splice(this.first, this.last - this.first + 1);
-  }
-}
-
-class ConditionGroup {
-  conditions;
-
-  constructor(conditions) {
-    if (!Array.isArray(conditions) || conditions.length < 2) {
-      throw Error("Cannot construct a condition group from a single condition");
-    }
-    this.conditions = conditions;
-  }
-
-  coordinatorString() {
-    return this.conditions[0].coordinatorString();
-  }
-
-  conditionString() {
-    const copy = [...this.conditions];
-    copy.splice(0, 1);
-    return `(${this.conditions[0].conditionString()} ${copy
-      .map((condition) => toPresentationString(condition))
-      .join(" ")})`;
-  }
-
-  conditionExpression() {
-    const copy = [...this.conditions];
-    copy.splice(0, 1);
-    return `(${this.conditions[0].conditionExpression()} ${copy
-      .map((condition) => toExpression(condition))
-      .join(" ")})`;
-  }
-
-  asFirstCondition() {
-    this.conditions[0].asFirstCondition();
-    return this;
-  }
-
-  getCoordinator() {
-    return this.conditions[0].getCoordinator();
-  }
-
-  setCoordinator(coordinator) {
-    this.conditions[0].setCoordinator(coordinator);
-  }
-
-  isGroup() {
-    return true;
-  }
-
-  getGroupedConditions() {
-    return this.conditions.map((condition) => condition.clone());
-  }
-
-  clone() {
-    return new ConditionGroup(
-      this.conditions.map((condition) => condition.clone())
-    );
-  }
-}
-
-export function toPresentationString(condition) {
-  return `${condition.coordinatorString()}${condition.conditionString()}`;
-}
-
-export function toExpression(condition) {
-  return `${condition.coordinatorString()}${condition.conditionExpression()}`;
-}
-
-export class Field {
-  name;
-  type;
-  display;
-
-  constructor(name, type, display) {
-    if (!name || typeof name !== "string") {
-      throw Error(`name ${name} is not valid`);
-    }
-    if (!ComponentTypes.find((componentType) => componentType.name === type)) {
-      throw Error(`type ${type} is not valid`);
-    }
-    if (!display || typeof display !== "string") {
-      throw Error(`display ${display} is not valid`);
-    }
-    this.name = name;
-    this.type = type;
-    this.display = display;
-  }
-
-  static from(obj) {
-    return new Field(obj.name, obj.type, obj.display);
-  }
-}
-
-class AbstractCondition {
-  coordinator;
-
-  constructor(coordinator) {
-    if (coordinator && !Object.values(coordinators).includes(coordinator)) {
-      throw Error(`coordinator ${coordinator} is not a valid coordinator`);
-    }
-    this.coordinator = coordinator;
-  }
-
-  coordinatorString() {
-    return this.coordinator ? `${this.coordinator} ` : "";
-  }
-
-  getCoordinator() {
-    return this.coordinator;
-  }
-
-  setCoordinator(coordinator) {
-    this.coordinator = coordinator;
-  }
-
-  isGroup() {
-    return false;
-  }
-
-  getGroupedConditions() {
-    return [this];
-  }
-
-  _asFirstCondition() {
-    delete this.coordinator;
-  }
-
-  asFirstCondition() {
-    throw Error(
-      "Implement on the subclass (Why do we have to have this method here at all?!)"
-    );
-  }
-
-  clone() {
-    throw Error(
-      "Implement on the subclass (Why do we have to have this method here at all?!)"
-    );
-  }
-
-  conditionString() {
-    throw Error(
-      "Implement on the subclass (Why do we have to have this method here at all?!)"
-    );
-  }
-
-  conditionExpression() {
-    throw Error(
-      "Implement on the subclass (Why do we have to have this method here at all?!)"
-    );
-  }
-}
-
-export class Condition extends AbstractCondition {
-  field;
-  operator;
-  value;
-
-  constructor(field, operator, value, coordinator) {
-    super(coordinator);
-    if (!(field instanceof Field)) {
-      throw Error(`field ${field} is not a valid Field object`);
-    }
-    if (typeof operator !== "string") {
-      throw Error(`operator ${operator} is not a valid operator`);
-    }
-    if (!(value instanceof AbstractConditionValue)) {
-      throw Error(`value ${value} is not a valid value type`);
-    }
-    this.field = field;
-    this.operator = operator;
-    this.value = value;
-  }
-
-  asFirstCondition() {
-    this._asFirstCondition();
-    return this;
-  }
-
-  conditionString() {
-    return `'${this.field.display}' ${
-      this.operator
-    } '${this.value.toPresentationString()}'`;
-  }
-
-  conditionExpression() {
-    return getExpression(
-      this.field.type,
-      this.field.name,
-      this.operator,
-      this.value
-    );
-  }
-
-  clone() {
-    return new Condition(
-      Field.from(this.field),
-      this.operator,
-      this.value.clone(),
-      this.coordinator
-    );
-  }
-}
-
-export class ConditionRef extends AbstractCondition {
-  conditionName;
-  conditionDisplayName;
-
-  constructor(conditionName, conditionDisplayName, coordinator) {
-    super(coordinator);
-    if (typeof conditionName !== "string") {
-      throw Error(`condition name ${conditionName} is not valid`);
-    }
-    if (typeof conditionDisplayName !== "string") {
-      throw Error(
-        `condition display name ${conditionDisplayName} is not valid`
-      );
-    }
-    this.conditionName = conditionName;
-    this.conditionDisplayName = conditionDisplayName;
-  }
-
-  asFirstCondition() {
-    this._asFirstCondition();
-    return this;
-  }
-
-  conditionString() {
-    return `'${this.conditionDisplayName}'`;
-  }
-
-  conditionExpression() {
-    return this.conditionName;
-  }
-
-  clone() {
-    return new ConditionRef(
-      this.conditionName,
-      this.conditionDisplayName,
-      this.coordinator
-    );
-  }
 }
