+++ model/src/conditions/inline-condition-operators.ts
@@ -1,3 +1,6 @@
+import { ComponentName } from "../component-types";
+import { AbstractConditionValue } from "./inline-condition-values";
+
 import {
   ConditionValue,
   dateDirections,
@@ -81,7 +84,12 @@ export function getOperatorNames(fieldType) {
   return Object.keys(getConditionals(fieldType)).sort();
 }
 
-export function getExpression(fieldType, fieldName, operator, value) {
+export function getExpression(
+  fieldType: ComponentName,
+  fieldName: string,
+  operator: string,
+  value: AbstractConditionValue
+) {
   return getConditionals(fieldType)[operator].expression(
     { type: fieldType, name: fieldName },
     value
@@ -92,7 +100,7 @@ export function getOperatorConfig(fieldType, operator) {
   return getConditionals(fieldType)[operator];
 }
 
-function getConditionals(fieldType) {
+function getConditionals(fieldType: ComponentName) {
   return customOperators[fieldType] || defaultOperators;
 }
